Removed Empty file, moved Composable to its own file
