--- app/src/main/java/com/example/randomuser/components/CustomIcon.kt
@@ -1,10 +1,0 @@
-package com.example.randomuser.components
-
-/*
-@Composable
-fun CustomIcon() {
-
-    Icon(imageVector = Icons.Filled.Male, contentDescription = )
-
-   // Icon(icon!!, descr, androidx.compose.ui.Modifier.size(iconSize), tint = iconTint!!)
-}*/
